feat: 删掉src，按照阿里云的规范
