+++ lib/helper.js
@@ -1,10 +1,10 @@
 'use strict';
 
-var xml2js = require('xml2js');
+const xml2js = require('xml2js');
 
 exports.parseXML = function (input) {
-  return new Promise(function (resolve, reject) {
-    xml2js.parseString(input, function (err, obj) {
+  return new Promise((resolve, reject) => {
+    xml2js.parseString(input, (err, obj) => {
       if (err) {
         return reject(err);
       }
@@ -13,13 +13,13 @@ exports.parseXML = function (input) {
   });
 };
 
-exports.extract = function extract(arr) {
+exports.extract = function extract (arr) {
   if (arr && arr.length === 1 && typeof arr[0] === 'string') {
     return arr[0];
   }
 
-  arr.forEach(function (item) {
-    Object.keys(item).forEach(function (key) {
+  arr.forEach((item) => {
+    Object.keys(item).forEach((key) => {
       item[key] = extract(item[key]);
     });
   });
@@ -27,14 +27,14 @@ exports.extract = function extract(arr) {
   return arr;
 };
 
-function format(params) {
+function format (params) {
   var xml = '';
-  Object.keys(params).forEach(function (key) {
-    var value = params[key];
+  Object.keys(params).forEach((key) => {
+    const value = params[key];
     if (typeof value === 'object') {
-      xml += `<${key}>${format(value)}</${key}>`;
+      xml +=    `<${key}>${format(value)}</${key}>`;
     } else {
-      xml += `<${key}>${value}</${key}>`;
+      xml +=    `<${key}>${value}</${key}>`;
     }
   });
   return xml;
@@ -42,17 +42,17 @@ function format(params) {
 
 exports.toXMLBuffer = function (entityType, params, subType) {
   var xml = '<?xml version="1.0" encoding="UTF-8"?>';
-  xml += `<${entityType} xmlns="http://mns.aliyuncs.com/doc/v1/">`;
+  xml +=    `<${entityType} xmlns="http://mns.aliyuncs.com/doc/v1/">`;
   if (Array.isArray(params)) {
-    params.forEach(function (item) {
-      xml += `<${subType}>`;
+    params.forEach((item) => {
+      xml +=  `<${subType}>`;
       xml += format(item);
-      xml += `</${subType}>`;
+      xml +=  `</${subType}>`;
     });
   } else {
-    xml += format(params);
+    xml +=    format(params);
   }
-  xml += `</${entityType}>`;
+  xml +=    `</${entityType}>`;
   return Buffer.from(xml, 'utf8');
 };
 
@@ -61,8 +61,8 @@ exports.toXMLBuffer = function (entityType, params, subType) {
 // http://{AccountId}.mns.cn-beijing-internal-vpc.aliyuncs.com
 
 exports.getEndpoint = function (accountid, opts) {
-  var protocol = opts.secure ? 'https' : 'http';
-  var region = `${opts.region}`;
+  const protocol = opts.secure ? 'https' : 'http';
+  let region = `${opts.region}`;
   if (opts.internal) {
     region += '-internal';
   }
@@ -78,16 +78,16 @@ exports.getEndpoint = function (accountid, opts) {
 };
 
 exports.getCanonicalizedMNSHeaders = function (headers) {
-  return Object.keys(headers).filter(function (key) {
-    return key.startsWith('x-mns-');
-  }).sort().map(function (key) {
-    return `${key}:${headers[key]}\n`;
-  }).join('');
+  return Object.keys(headers)
+    .filter((key) => key.startsWith('x-mns-'))
+    .sort()
+    .map((key) => `${key}:${headers[key]}\n`)
+    .join('');
 };
 
 exports.getResponseHeaders = function (headers, attentions) {
   var result = {};
-  attentions.forEach(function (key) {
+  attentions.forEach((key) => {
     result[key] = headers[key];
   });
   return result;
